Simple Tile Layout and Character Created
Completed the creation of a test level with a player and an area to test navigation on,
the player can move and collide with the walls in the level.
Next phase is to add a navigation mesh and attempt to create  display pathfinding from the
player position to the desired location.