Working Enemy Pathfinding
Managed to add enemies and have them use the navmesh to move towards a target, the main issue
however is with the physics collisions. When allowed to collide with each other they are  too
regularly clipping each other into walls or off the navmesh. This will require a more specialized
collision handler. That aside  the code is running well.